Add migration for papers and footnotes
This commit adds a migration for papers, and their footnotes. This 
migration is very similar to the migrations we have in rails, where 'up'
is adding to our database and 'down' is removing.

In both table creations, we make an id column that auto-increments
and is a primary key (noted by the primary() function chained to it).

We create our other columns with recognizable types (integer, string).

On the integer field for paper_id in the footnotes table, we use the
unsigned() function to make this column an unsigned integer, meaning
we have no use for the negative numbers that integer normally provides,
reserving more memory for a larger positive number, since foreign keys
are never negative anyway.

Finally, we tell knex that paper_id is a foreign key column and that
it references papers.id.