Add server file for express with configuration
This file adds the configuration for a small and simple Express server.
At the top, we require our Express module and start an initiation of its
function express(). We use whatever environment is set but default to
development if none is found. We use the bodyParser.json to get the 
ability to parse JSON and Urls as well with the next line.

Function on line 14 adds the ability to have a root path at '/'. The
next function on line 18 lets us always listen on the port for a 
connection request.