Add route actions for papers and footnotes
This commit is quite large but adds a couple of things we are already
familiar with from Rails/Sinatra.

Lines 18 and 28 have the functions for the get action for all papers
and footnotes. The respective tables are selected via the select
function, and then returned as JSON to the client that made the request.

Line 38's function is the POST action for adding a new paper to our
database. The body is assigned to the const paper. The for loop below
it iterates over the column names we would like to be required for a
successful POST action to occur. If one of the fields is missing,
we return an error and stop the action and let the user know which
field is missing. Otherwise, the paper is inserted into the database.

On line 49 where the insertion happens, the second argument in the insert
function is what we would like returned once the insertion happens. We
then respond with a 201 content created status with the JSON going back
out to the client.

The function on line 58 adds a path for inserting a footnote into the 
database based on a given paper id. We went this route because it ensures 
that a footnote is only added to a paper with a valid paper id (since
we have added paper_id as a foreign_key in our migration).

Line 79 has our last action for this server. Its a GET request for a
specific paper given by the :id in the URI. It attempts to find the 
paper in the database given the specific ID. If it cannot find it, it
returns the error back to the client with a 404 code. Otherwise, it
retrieves the valid paper from the database and returns it as JSON to
the user.